refactor: restructure repository for production-ready architecture
