--- app/api/__init__.py
@@ -1,4 +1,0 @@
-from app.api.repositories import router as repositories_router
-from app.api.chat import router as chat_router
-
-__all__ = ["repositories_router", "chat_router"]
