+++ backend/api/__init__.py
@@ -1,0 +1,4 @@
+from backend.api.repositories import router as repositories_router
+from backend.api.chat import router as chat_router
+
+__all__ = ["repositories_router", "chat_router"]
