+++ backend/middleware/__init__.py
@@ -4,8 +4,8 @@ Middleware Module
 Phase 13: Production Ready
 """
 
-from app.middleware.rate_limiter import limiter, rate_limit
-from app.middleware.error_handlers import register_error_handlers
+from backend.middleware.rate_limiter import limiter, rate_limit
+from backend.middleware.error_handlers import register_error_handlers
 
 __all__ = [
     "limiter",
