+++ backend/services/__init__.py
@@ -1,0 +1,28 @@
+from backend.services.file_scanner import scan_directory
+from backend.services.ast_parser import CodeExtractor, parse_python_file
+from backend.services.chunker import CodeChunker
+from backend.services.embedding import EmbeddingService
+from backend.services.vector_store import VectorStore
+from backend.services.keyword_search import KeywordSearchService
+from backend.services.hybrid_search import HybridSearchService
+from backend.services.llm_service import LLMService
+from backend.services.processor import RepositoryProcessor
+from backend.services.rag_pipeline import RAGPipeline
+from backend.services.github_service import GitHubService
+from backend.services.chat_service import ChatService
+
+__all__ = [
+    "scan_directory",
+    "CodeExtractor",
+    "parse_python_file",
+    "CodeChunker",
+    "EmbeddingService",
+    "VectorStore",
+    "KeywordSearchService",
+    "HybridSearchService",
+    "LLMService",
+    "RepositoryProcessor",
+    "RAGPipeline",
+    "GitHubService",
+    "ChatService",
+]
