+++ backend/services/file_scanner.py
