+++ backend/services/keyword_search.py
@@ -7,7 +7,7 @@ Integrates with the RAG pipeline for hybrid search.
 Phase 7: Keyword Search with MongoDB
 """
 
-from app.database import Database
+from backend.database import Database
 from typing import List, Dict, Any, Optional
 import re
